requirements.txt
easyocr
pandas
numpy
requests
rapidfuzz
streamlit
python-dotenv